% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_fit.R
\name{copula_fit}
\alias{copula_fit}
\title{Fit copula to uniform residuals.}
\usage{
copula_fit(spec, u, X = NULL, distribution = "norm", constant = TRUE,
  upsilon = FALSE)
}
\arguments{
\item{spec}{CopulaSpecification to fit}

\item{u}{TxN uniform residuals}

\item{constant}{Optimize constant copula (alpha = beta = 0)}

\item{NxkxT}{array of external regressors}
}
\value{
Fitted CopulaSpecification
}
\description{
Note: This takes a long time.
}

