% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_fit.R
\name{copula_fit_pars}
\alias{copula_fit_pars}
\title{Get pars and constraints for a copula model}
\usage{
copula_fit_pars(spec, distribution = "norm", constant = T, upsilon = F)
}
\arguments{
\item{distribution}{norm, t or ghst}

\item{constant}{constant copula?}

\item{upsilon}{with upsilon?}

\item{N}{how many factors?}
}
\value{
list with theta, ui and ci
}
\description{
Get pars and constraints for a copula model
}

