% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_fit.R
\name{copula_fit_theta}
\alias{copula_fit_theta}
\title{Get theta and constraints for a copula model}
\usage{
copula_fit_theta(N, distribution = "norm", constant = T, upsilon = F)
}
\arguments{
\item{N}{how many factors?}

\item{distribution}{norm, t or ghst}

\item{constant}{constant copula?}

\item{upsilon}{with upsilon?}
}
\value{
list with theta, ui and ci
}
\description{
Get theta and constraints for a copula model
}

