% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copula_Upsilon}
\alias{copula_Upsilon}
\title{Build Upsilon array from Nxk coefficients and independent variables}
\usage{
copula_Upsilon(theta, X)
}
\arguments{
\item{theta}{Nxk matrix of regressors}

\item{X}{Nxkxt array of independent data}
}
\value{
Nxkxt Upsilon array (correlation matrix)
}
\description{
It's impossible for the copula to know how many of your theta are "free"
or how many of your X are identical; it just takes the Upsilon series as
given. You can construct it yourself, or use this function to build an
appropriate form following the appendix in Christoffersen et al. (2012).
}

