% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{garch_copula_filter}
\alias{garch_copula_filter}
\title{Filter a return series through GARCH Copula Model}
\usage{
garch_copula_filter(model, x, X = NULL)
}
\arguments{
\item{model}{GARCHCopulaModel}

\item{x}{Return series}

\item{X}{Upsilon regressors}
}
\value{
result of copula_filter
}
\description{
Filter a return series through GARCH Copula Model
}

