% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch_fit}
\alias{garch_fit}
\title{Fit GARCH models to data}
\usage{
garch_fit(garch, data)
}
\arguments{
\item{garch}{List of ugarchspec}

\item{data}{TxN data points}
}
\description{
Fit GARCH models to data
}

