% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch_fit2spec}
\alias{garch_fit2spec}
\title{Turns GARCH fits to specs with fixed.pars}
\usage{
garch_fit2spec(fits)
}
\arguments{
\item{fits}{List of GARCH fits}
}
\value{
List of GARCH specs
}
\description{
Turns GARCH fits to specs with fixed.pars
}

