% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch_specgen}
\alias{garch_specgen}
\title{Create GARCH specifications}
\usage{
garch_specgen(p, q = 0, r = 1, s = 1, model = "fGARCH",
  submodel = "GJRGARCH", vtarget = T, dist = "ghst",
  fixed.pars = list())
}
\arguments{
\item{p}{Scalar - AR order}

\item{q}{Scalar - MA order}

\item{r}{Scalar - GARCH alpha order}

\item{s}{Scalar - GARCH beta order}

\item{vtarget}{Logical - setting long-term variance using unconditional mean}
}
\value{
uGARCHspec object
}
\description{
Makes building GARCH models easier as standard options
apart from p, q order is preset. Standard options include
variance targeting, GJRGACH(1,1) specification and the GHST
distribution for innovations
}

