% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch_stdresid2uniform}
\alias{garch_stdresid2uniform}
\title{Turn standardized residuals to uniforms. Only works with ghst GARCH}
\usage{
garch_stdresid2uniform(garch, stdresid)
}
\arguments{
\item{garch}{N list of GARCH models}

\item{stdresid}{TxN matrix of stdresid}
}
\description{
Turn standardized residuals to uniforms. Only works with ghst GARCH
}

