% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch_uniform2stdresid}
\alias{garch_uniform2stdresid}
\title{Turn uniform to stdresid for a list of GARCH models}
\usage{
garch_uniform2stdresid(garch, u)
}
\arguments{
\item{garch}{N list of GARCH models}

\item{u}{TxN uniforms}
}
\description{
Turn uniform to stdresid for a list of GARCH models
}

