% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copula_simulate}
\alias{copula_simulate}
\title{Simulate uniform residuals from the copula}
\usage{
copula_simulate(spec, n.sim, m.sim, Q_T = NULL, shocks_T = NULL, X = NULL)
}
\arguments{
\item{spec}{CopulaSpecification}

\item{n.sim}{Simulation horizon}

\item{m.sim}{Number of simulation}

\item{Q_T}{The final Q_t your dataset (default: Omega)}

\item{shocks_T}{Final shocks in dataset}

\item{X}{Nxkxn.sim exogenous regressors}
}
\value{
List of m.sim n.simxN uniform residuals
}
\description{
Simulate uniform residuals from the copula
}

